+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ###
 1. Added github nightly build with executable artifacts.
+2. Changed default controller timezone from UTC to Local.
 
 
 ## [0.8.8](https://github.com/uhppoted/uhppote-cli/releases/tag/v0.8.8) - 2024-03-27